+++ manual/content/appendices/release-notes.md
@@ -10,9 +10,15 @@ This section includes release notes for the Syntoniq software. See also [docs/TO
 
 # v0.4.1 - April 2, 2026
 
+## Enhancements
+
 * Add the `csound_global_instrument` directive for enabling global effect instruments, such as reverb, when using custom Csound templates.
 * Add the `csound_template` directive, which takes a relative path, for specifying a Csound template. This can still be overridden from the command line.
 * Add prompt syntax for saving pitches to and restoring pitches from variables.
+
+## Bug Fixes
+
+* Remove false "unknown part" error in Csound processor dynamics lines that appear in parts that don't otherwise appear in the region.
 
 # v0.4.0 - April 2, 2026